Add default IQ point and make type public
The type IQData.Point was previously internal which messed up the pretty printer
in printfn for no benefit, so now it's a public type.

Also added a default IQ point record, so the standard method of construction
ought now to be by piping that through the "withIPoint" etc. family of
functions.