Add byte array intepretation to agent
Data must be sent as a byte array rather than F#'s internal UTF-8 string
representation as IQ data could have the high bit set, which would cause one byte
to become multiple bytes if it was encoded to UTF-8, destroying the data.  The
IVisa interface needed to become aware of this, and have functions to handle the
byte arrays.

Updated the dummy instrument to respond to the new abstract members added,
though it doesn't offer anything by way of replies.