Add deletion of sequences and update WaveformDemo
Segments and sequences can now both be written and deleted from the machine with
no problems.

Also updated WaveformDemo to give a more accurate demonstration of what the
system is currently capable of.

Did some tidying up of the codebase to make it more legible and less like a
testing product.