Make Verify module for RfPulses
Separated out a verification step in the generation of an Rf pulse experiment so
that the semantics of the compile and compress steps can make assumptions.  We
can return a Failure if verification fails, meaning that we don't need to go
through the process of compilation and compression if there is an error in the
input - some things can't be checked by the type system, like the lengths of
lists generated at runtime.

Also moved the Compile and Compress steps into their own modules to make the
whole module flow nicer, and to clear up the number of helper functions which
are available, even in private scope.