Fix PicoScope.ChannelSettings.queryAvailableChannels implementation to account for power source
Updated the queryAvailableChannels function to also check the current power source and take this into account when querying the channels available with the current settings. However, streaming acquisition still fails at when the call to RunStreaming is made. It appears that it the device is unable to perform streaming acquisition with USB power. Therefore this closes #44.