Rewrite VISA agent using exceptions and new IVI VISA library
This new agent is very similar in style to the old one, but uses exception
handling instead of AsyncChoice to deal with errors.  Internally in the agent,
we still use Choice, but instead of string error messages, we now use
exceptions.  These choice failures are propagated up until they are outside the
agent, when we either raise the exception, or return the value.  This allows the
agent to continue operating, even on failure.

The version of the VISA library was also upgraded to 15.0, from the IVI
Foundation, rather than the previous NationalInstruments version.  Hopefully,
this should make it a little more future-proof (particularly since we're using
.NET), but I'm not holding my breath.