Update VISA agent to include an optional delay after writing commands to the hardware
The writeDelay parameter is specified when opening the VISA instrument, so the interface for opening instruments is updated. Furthermore, the interface for closing instruments is also updated as it followed the outdated AsyncChoice-based error handling architecture. Therefore, usage of the VISA agent in the Keysight RF source and LakeShore temperature controller projects is also updated.