Add synchronous I/O functions to VISA agent and add protocol-specific openInstrument variants
Serial port (RS-232) instruments require parameters such as parity and stop bits to be set correctly in order to function so a new type has been introduced to describe the serial configuration. However, this is not required for GPIB and Ethernet instruments so now there are protocol-specific openInstrument functions and openSerialInstrument requires the serial configuration.

Secondly, because of the VISA library behaviour when performing asynchronous I/O, this can lead to a bottleneck when sending many short commands. This is because additional messages need to be sent in order to fascilitate asynchronous I/O. The advantage of asynchronous I/O is that the agent thread will not hang during the I/O operation. Therefore, functionality has been introduced to also perform synchronous alongside the asynchronous ones.

VISA agent usage has been updated accordingly in the existing instrument libraries, defaulting to asynchronous I/O.