Add SCPI library
The library can build up SCPI commands, query the error queue, build SCPI
datablocks and identify SCPI machines.  It can return ASCII strings of these
commands, or it can write to instruments which implement the defined
SCPI.IScpiInstrument interface functions, with an optional terminator.

Everything is done as ASCII strings, because technically that's a requirement of
the SCPI spec.  We exploit the fact that 7-bit ASCII is entirely compatible with
the UTF-8 encoding of .NET strings, so SCPI keys can be passed as standard .NET
strings, without the user having to call an encoding converter, and we maintain
low-level interoperability with C.

Generic setters and queriers are implemented, where values to be written are
converted to strings using their ToString() methods.  I chose this over using a
"mapToString" passed argument because overall it makes for much nicer semantics,
and the majority of types we expect to be passed (int, etc), already have nice
ToString() methods.  Derived types can simply override these in the instrument
drivers, or can manually value map the values into strings.